requirements.txt
feedparser==6.0.11
requests==2.31.0
beautifulsoup4==4.12.3
langdetect==1.0.9
python-dotenv==1.0.1
